requirements.txt
torch
transformers
gradio
pillow